Typecast fixed.

+++ src/Socket.cpp
@@ -99,7 +99,7 @@ namespace eth
         device.executeSocketCommand(m_handle, SocketCommand::send);
 
 
-        constexpr uint8_t send = static_cast<int>(SocketInterrupt::send);
+        constexpr uint8_t send = static_cast<uint8_t>(SocketInterrupt::send);
 
         while( ( device.readSocketInterruptRegister(m_handle) & send ) != send )
         {
